add pasta routes e assets
