+++ labecommerce/src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Router from './Router';
+import Router from './routes/Router';
 
 function App() {
   return (
