+++ labecommerce/src/routes/Coordinator.js
@@ -1,0 +1,7 @@
+export const goToHome = (navigate) => {
+    navigate("/")
+  }
+
+  export const goToCarrinho = (navigate) => {
+    navigate("/carrinho")
+  }
